Refactor repository update logic into separate class
The repository update logic previously found in MobileRefreshPage and RefreshPage has been moved to a new class, RepositoryUpdater. This refactor improves the code structure by separating concerns, making the logic reusable and simplifying the parent classes.
