+++ app/src/main/java/com/dlsc/jfxcentral2/app/pages/RefreshPage.java
@@ -1,7 +1,7 @@
 package com.dlsc.jfxcentral2.app.pages;
 
-import com.dlsc.jfxcentral.data.DataRepository2;
 import com.dlsc.jfxcentral2.app.RepositoryManager;
+import com.dlsc.jfxcentral2.app.utils.RepositoryUpdater;
 import com.dlsc.jfxcentral2.components.CustomImageView;
 import com.dlsc.jfxcentral2.components.Mode;
 import com.dlsc.jfxcentral2.model.Size;
@@ -9,11 +9,7 @@ import javafx.application.Platform;
 import javafx.beans.InvalidationListener;
 import javafx.beans.WeakInvalidationListener;
 import javafx.beans.binding.Bindings;
-import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.SimpleIntegerProperty;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
 import javafx.scene.control.Button;
@@ -24,7 +20,6 @@ import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
 import javafx.scene.text.TextAlignment;
 import one.jpro.platform.routing.sessionmanager.SessionManager;
-import org.eclipse.jgit.lib.ProgressMonitor;
 
 public class RefreshPage extends PageBase {
 
@@ -44,6 +39,7 @@ public class RefreshPage extends PageBase {
     };
 
     private final WeakInvalidationListener weakInvalidationListener = new WeakInvalidationListener(invalidationListener);
+    private final RepositoryUpdater repositoryUpdater = new RepositoryUpdater();
 
     private Node updateView;
     private Node setupView;
@@ -61,10 +57,6 @@ public class RefreshPage extends PageBase {
     public String description() {
         return "A central place for anything related to JavaFX.";
     }
-
-    private final IntegerProperty loadPercentage = new SimpleIntegerProperty(this, "loadPercentage", 0);
-
-    private final StringProperty loadMessage = new SimpleStringProperty(this, "loadMessage", "");
 
     @Override
     public boolean fullscreen() {
@@ -87,7 +79,7 @@ public class RefreshPage extends PageBase {
         setupView.setVisible(RepositoryManager.isFirstTimeSetup());
 
         if (!RepositoryManager.isFirstTimeSetup()) {
-            performUpdate(true);
+            repositoryUpdater.performUpdate(true);
         }
 
         return new StackPane(updateView, setupView);
@@ -111,7 +103,7 @@ public class RefreshPage extends PageBase {
             updateView.setManaged(true);
             setupView.setVisible(false);
             setupView.setManaged(false);
-            performUpdate(false);
+            repositoryUpdater.performUpdate(false);
         });
 
         Button exitButton = new Button("EXIT");
@@ -141,16 +133,16 @@ public class RefreshPage extends PageBase {
         Label label = new Label();
         label.setTextAlignment(TextAlignment.CENTER);
         label.textProperty().bind(Bindings.createStringBinding(() -> {
-            if (loadMessage.get().toLowerCase().startsWith("pull")) {
+            if (repositoryUpdater.getLoadMessage().toLowerCase().startsWith("pull")) {
                 return "Checking for updates ...";
             }
 
-            if (loadPercentage.get() == -1) {
-                return loadMessage.get();
+            if (repositoryUpdater.getLoadPercentage() == -1) {
+                return repositoryUpdater.getLoadMessage();
             }
 
-            return loadMessage.get() + " " + loadPercentage.get() + "%";
-        }, loadMessage, loadPercentage));
+            return repositoryUpdater.getLoadMessage() + " " + repositoryUpdater.getLoadPercentage() + "%";
+        }, repositoryUpdater.loadMessageProperty(), repositoryUpdater.loadPercentageProperty()));
         label.setMaxSize(Double.MAX_VALUE, Double.MAX_VALUE);
 
         VBox vBox = new VBox(logo, label);
@@ -161,72 +153,5 @@ public class RefreshPage extends PageBase {
         wrapper.getStyleClass().add("update-view");
 
         return wrapper;
-    }
-
-    /*
-     * The "refresh" parameter states whether the update is called due to the first setup of the
-     * data repository or because of a call to the refresh page.
-     */
-    private void performUpdate(boolean refresh) {
-        Thread thread = new Thread(() -> RepositoryManager.updateRepository(new ProgressMonitor() {
-
-            double total;
-            double acc;
-
-            @Override
-            public void showDuration(boolean b) {
-            }
-
-            @Override
-            public void start(int i) {
-            }
-
-            @Override
-            public void beginTask(String taskName, int totalWork) {
-                // clean up some messy message we get
-                final String name = taskName.replace("remote: ", "");
-
-                if (totalWork == -1) {
-                    Platform.runLater(() -> {
-                        loadMessage.set(name);
-                        loadPercentage.set(-1);
-                    });
-                } else {
-                    Platform.runLater(() -> {
-                        loadMessage.set(name);
-                        loadPercentage.set(0);
-                    });
-                    acc = 0;
-                    total = totalWork;
-                }
-            }
-
-            @Override
-            public void update(int completed) {
-                if (total != ProgressMonitor.UNKNOWN) {
-                    acc += completed;
-                    int p = (int) ((acc / total) * 100);
-                    Platform.runLater(() -> loadPercentage.set(p));
-                } else {
-                    Platform.runLater(() -> loadPercentage.set(-1));
-                }
-            }
-
-            @Override
-            public void endTask() {
-                if (refresh) {
-                    DataRepository2.getInstance().reload();
-                }
-            }
-
-            @Override
-            public boolean isCancelled() {
-                return false;
-            }
-        }));
-
-        thread.setName("Repository Thread");
-        thread.setDaemon(true);
-        thread.start();
     }
 }
